+++ threedi_modelchecker/tests/test_checks_base.py
@@ -603,8 +603,8 @@ def test_list_of_inst_check(session, tag_ids_string, nof_invalid_expected):
 )
 def test_epsg_geom_check(session, ref_epsg, valid):
     factories.ConnectionNodeFactory()
-    session.ref_epsg_code = ref_epsg
-    session.ref_epsg_name = "foo"
+    session.epsg_ref_code = ref_epsg
+    session.epsg_ref_name = "foo"
     check = EPSGGeomCheck(column=models.ConnectionNode.geom)
     invalids = check.get_invalid(session)
     assert (len(invalids) == 0) == valid