+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog of threedi-modelchecker
 2.17.1 (unreleased)
 -------------------
 
+- Fix typos in session.epsg_ref_name and session.epsg_ref_code
 - Ensure that id's created via factory models, used for testing, always autoincrement
-
 
 
 2.17.0 (2025-01-24)